Rename Stretched Fixed Mode to Fixed Mode plugin
As this plugin is going to be extended with more features not relevant
to stretching of the fixed mode, rename the plugin.
